+++ extensions/src/extensions/getPage.ts
@@ -1,4 +1,4 @@
-import {GridFunction, Paging, PAGING} from '@anglr/grid';
+import {GridFunction, GridPluginType, Paging} from '@anglr/grid';
 
 /**
  * Gets current page of grid
@@ -7,7 +7,7 @@ export function getPage(): GridFunction<number>
 {
     return grid =>
     {
-        const paging = grid.getPlugin<Paging>(PAGING);
+        const paging = grid.getPlugin<Paging>(GridPluginType.Paging);
 
         return paging.page;
     };