+++ extensions/src/extensions/getSelectedData.ts
@@ -1,4 +1,4 @@
-import {GridFunction, RowSelector, ROW_SELECTOR} from '@anglr/grid';
+import {GridFunction, GridPluginType, RowSelector} from '@anglr/grid';
 
 /**
  * Gets currently selected data
@@ -7,7 +7,7 @@ export function getSelectedData<TSelectedData>(): GridFunction<TSelectedData[]>
 {
     return grid =>
     {
-        const rowSelector = grid.getPlugin<RowSelector<TSelectedData>>(ROW_SELECTOR);
+        const rowSelector = grid.getPlugin<RowSelector<TSelectedData>>(GridPluginType.RowSelector);
 
         return rowSelector.selectedData;
     };