+++ extensions/src/extensions/invalidateBodyContent.ts
@@ -1,4 +1,4 @@
-import {GridAction, BODY_CONTENT_RENDERER} from '@anglr/grid';
+import {GridAction, GridPluginType} from '@anglr/grid';
 
 /**
  * Invalidates body content renderer view
@@ -7,7 +7,7 @@ export function invalidateBodyContent(): GridAction
 {
     return grid =>
     {
-        const bodyContentRenderer = grid.getPlugin(BODY_CONTENT_RENDERER);
+        const bodyContentRenderer = grid.getPlugin('BODY_CONTENT_RENDERER' as unknown as GridPluginType);
 
         bodyContentRenderer.invalidateVisuals();
     };