+++ extensions/src/extensions/resetSelection.ts
@@ -1,4 +1,4 @@
-import {GridAction, RowSelector, ROW_SELECTOR} from '@anglr/grid';
+import {GridAction, GridPluginType, RowSelector} from '@anglr/grid';
 
 /**
  * Resets (deselects) all selection
@@ -7,7 +7,7 @@ export function resetSelection(): GridAction
 {
     return grid =>
     {
-        const rowSelector = grid.getPlugin<RowSelector>(ROW_SELECTOR);
+        const rowSelector = grid.getPlugin<RowSelector>(GridPluginType.RowSelector);
 
         rowSelector.resetSelection();
     };