+++ extensions/src/extensions/setOrdering.ts
@@ -1,4 +1,4 @@
-import {ContentRenderer, CONTENT_RENDERER, DataLoader, DATA_LOADER, GridAction} from '@anglr/grid';
+import {ContentRenderer, DataLoader, GridAction, GridPluginType} from '@anglr/grid';
 
 /**
  * Sets ordering for grid
@@ -8,10 +8,10 @@ export function setOrdering<TOrdering>(ordering: TOrdering): GridAction
 {
     return grid =>
     {
-        const contentRenderer = grid.getPlugin<ContentRenderer<TOrdering>>(CONTENT_RENDERER);
-        const dataLoader = grid.getPlugin<DataLoader>(DATA_LOADER);
+        const contentRenderer = grid.getPlugin<ContentRenderer<TOrdering>>(GridPluginType.ContentRenderer);
+        const dataLoader = grid.getPlugin<DataLoader>(GridPluginType.DataLoader);
 
-        contentRenderer.ordering = ordering;
+        contentRenderer.ordering = ordering as TOrdering;
         dataLoader.loadData();
     };
 }