+++ extensions/src/extensions/setSyncData.ts
@@ -1,4 +1,4 @@
-import {GridAction, DataLoader, DATA_LOADER} from '@anglr/grid';
+import {GridAction, DataLoader, GridPluginType, SyncDataLoaderOptions} from '@anglr/grid';
 import {isArray} from '@jscrpt/common';
 
 /**
@@ -6,13 +6,13 @@ import {isArray} from '@jscrpt/common';
  * @param data - data that should be replaced
  * @param force - Indication that data should be reloaded 
  */
-export function setSyncData(data: any[], force?: boolean): GridAction
+export function setSyncData(data: unknown[], force?: boolean): GridAction
 {
     return grid =>
     {
-        const dataLoader = grid.getPlugin<DataLoader>(DATA_LOADER);
+        const dataLoader = grid.getPlugin<DataLoader>(GridPluginType.DataLoader);
 
-        dataLoader.options.data = isArray(data) ? data : [];
+        (dataLoader.options as SyncDataLoaderOptions).data = isArray(data) ? data : [];
         dataLoader.loadData(force);
     };
 }