+++ extensions/src/extensions/showMetadataSelector.ts
@@ -1,4 +1,4 @@
-import {GridAction, METADATA_SELECTOR, MetadataSelector} from '@anglr/grid';
+import {GridAction, GridPluginType, MetadataSelector} from '@anglr/grid';
 
 /**
  * Shows metadata (column) selector
@@ -7,7 +7,7 @@ export function showMetadataSelector(): GridAction
 {
     return grid =>
     {
-        const metadataSelector = grid.getPlugin<MetadataSelector>(METADATA_SELECTOR);
+        const metadataSelector = grid.getPlugin<MetadataSelector>(GridPluginType.MetadataSelector);
 
         metadataSelector.show();
     };