+++ src/plugins/dataLoader/async/asyncDataLoader.interface.ts
@@ -1,9 +1,9 @@
-import {DataLoaderOptions, DataLoader, DataResponse} from '../dataLoader.interface';
+import {DataLoader, DataLoaderOptions, DataResponse} from '../../../interfaces';
 
 /**
  * Async data loader options
  */
-export interface AsyncDataLoaderOptions<TData = any, TOrdering = any> extends DataLoaderOptions
+export interface AsyncDataLoaderOptions<TData = unknown, TOrdering = unknown> extends DataLoaderOptions
 {
     /**
      * Callback used for obtaining data
@@ -11,12 +11,12 @@ export interface AsyncDataLoaderOptions<TData = any, TOrdering = any> extends Da
      * @param itemsPerPage - Number of items per page
      * @param ordering - Order by column name
      */
-    dataCallback?: (page: number, itemsPerPage: number, ordering: TOrdering) => Promise<DataResponse<TData>>;
+    dataCallback: (page: number, itemsPerPage: number, ordering: TOrdering|undefined|null) => Promise<DataResponse<TData>>;
 }
 
 /**
  * Public API for 'AsyncDataLoaderComponent'
  */
-export interface AsyncDataLoader<TData = any> extends DataLoader<DataResponse<TData>>
+export interface AsyncDataLoader<TData = unknown> extends DataLoader<DataResponse<TData>>
 {
 }