+++ src/plugins/dataLoader/components.ts
@@ -1,3 +1,0 @@
-export * from './dataLoaderAbstract.component';
-export * from './async/asyncDataLoader.component';
-export * from './sync/syncDataLoader.component';