+++ src/plugins/dataLoader/index.ts
@@ -1,3 +1,5 @@
-export * from './dataLoader.interface';
+export * from './async/asyncDataLoader.component';
 export * from './async/asyncDataLoader.interface';
+export * from './dataLoaderAbstract.component';
+export * from './sync/syncDataLoader.component';
 export * from './sync/syncDataLoader.interface';