+++ src/plugins/dataLoader/sync/syncDataLoader.component.ts
@@ -1,26 +1,25 @@
 import {Component, ChangeDetectionStrategy, Inject, Optional, ElementRef} from '@angular/core';
-import {extend, OrderByDirection} from '@jscrpt/common';
+import {OrderByDirection} from '@jscrpt/common';
+import {lastValueFrom} from '@jscrpt/common/rxjs';
 import {from, Observable} from 'rxjs';
 import {skip, take, toArray} from 'rxjs/operators';
 
 import {SyncDataLoaderOptions, SyncDataLoader} from './syncDataLoader.interface';
-import {DATA_LOADER_OPTIONS, DataLoaderState} from '../types';
-import {GridPluginInstances} from '../../../components/grid';
-import {GRID_PLUGIN_INSTANCES} from '../../../components/grid/types';
-import {DataResponse} from '../dataLoader.interface';
 import {DataLoaderAbstractComponent} from '../dataLoaderAbstract.component';
-import {SimpleOrdering} from '../../contentRenderer';
+import {DataResponse, SimpleOrdering} from '../../../interfaces';
+import {DATA_LOADER_OPTIONS, GRID_PLUGIN_INSTANCES} from '../../../misc/tokens';
+import {GridPluginInstances} from '../../../misc/types';
+import {DataLoaderState} from '../../../misc/enums';
 
 /**
  * Default options for sync data loader
- * @internal
  */
-const defaultOptions: SyncDataLoaderOptions<any, SimpleOrdering> =
+const defaultOptions: SyncDataLoaderOptions<Record<string, number>, SimpleOrdering> =
 {
     autoLoadData: true,
     debounceDataCallback: 30,
     data: [],
-    orderData: (data: any[], ordering: SimpleOrdering) =>
+    orderData: (data: Record<string, number>[], ordering?: SimpleOrdering) =>
     {
         if(!ordering)
         {
@@ -46,45 +45,23 @@ const defaultOptions: SyncDataLoaderOptions<any, SimpleOrdering> =
 /**
  * Data loader that allows synchronous data loading
  *
- * @example
- * If you want to use this loader you have to provide all data you want to be displayed in grid
- *
- * You have to provide data at least
- * ``` typescript
- * var arrayOfItems: DataType[] = [];
- *
- * var gridOptions =
- * {
- *      plugins:
- *      {
- *          dataLoader:
- *          {
- *              type: SyncDataLoaderComponent,
- *              options: <SyncDataLoaderOptions<DataType, SimpleOrdering>>
- *              {
- *                  data: arrayOfItems
- *              }
- *          }
- *      }
- * };
- * ```
- * 
- * You must set these options before end of sync call of `ngOnInit`, if you set it later you have to disable auto initialization of grid and initialize it manualy
+ * You must set options before end of sync call of `ngOnInit`, if you set it later you have to disable auto initialization of grid and initialize it manualy
  */
 @Component(
 {
     selector: 'ng-sync-data-loader',
     template: '',
+    standalone: true,
     changeDetection: ChangeDetectionStrategy.OnPush
 })
-export class SyncDataLoaderComponent<TData = any, TOrdering = any> extends DataLoaderAbstractComponent<SyncDataLoaderOptions<TData, TOrdering>, TOrdering, DataResponse<TData>> implements SyncDataLoader<TData>
+export class SyncDataLoaderSAComponent<TData = unknown, TOrdering = unknown> extends DataLoaderAbstractComponent<SyncDataLoaderOptions<TData, TOrdering>, TOrdering, DataResponse<TData>> implements SyncDataLoader<TData>
 {
-    //######################### private fields #########################
+    //######################### protected fields #########################
 
     /**
      * Current result of data loader
      */
-    private _result?: DataResponse<TData> =
+    protected ɵresult: DataResponse<TData> =
     {
         data: [],
         totalCount: 0
@@ -93,61 +70,57 @@ export class SyncDataLoaderComponent<TData = any, TOrdering = any> extends DataL
     //######################### public properties #########################
 
     /**
-     * Current result of data loader
+     * @inheritdoc
      */
     public get result(): DataResponse<TData>
     {
-        return this._result;
+        return this.ɵresult;
     }
 
     //######################### constructor #########################
     constructor(pluginElement: ElementRef,
-                @Inject(GRID_PLUGIN_INSTANCES) @Optional() gridPlugins?: GridPluginInstances,
+                @Inject(GRID_PLUGIN_INSTANCES) @Optional() gridPlugins: GridPluginInstances|undefined|null,
                 @Inject(DATA_LOADER_OPTIONS) @Optional() options?: SyncDataLoaderOptions<TData, TOrdering>)
     {
-        super(pluginElement, gridPlugins);
-
-        this._options = extend(true, {}, defaultOptions, options);
+        super(pluginElement, gridPlugins, defaultOptions as unknown as SyncDataLoaderOptions<TData, TOrdering>, options);
     }
 
     //######################### protected methodes - implements DataLoaderAbstractComponent #########################
 
     /**
-     * Loads data from 'source'
-     * @param force - Indication that data should be reloaded even if nothing changed
+     * @inheritdoc
      */
-    protected async _loadData(force?: boolean)
+    protected async loadGridData(force?: boolean): Promise<void>
     {
-        if(!force && !this._checkChanges())
+        if(!force && !this.checkChanges())
         {
             return;
         }
 
-        let data = [...this._options.data];
+        let data = [...this.ɵoptions.data];
 
-        this._state = (data && data.length) ? DataLoaderState.DataLoading : DataLoaderState.NoDataLoading;
-        this.stateChange.emit();
+        this.ɵstate = (data && data.length) ? DataLoaderState.DataLoading : DataLoaderState.NoDataLoading;
+        this.stateChangeSubject.next();
 
-        if(this._options.orderData)
+        if(this.ɵoptions.orderData)
         {
-            data = this._options.orderData(data, this._contentRenderer.ordering);
+            data = this.ɵoptions.orderData(data, this.contentRenderer?.ordering);
         }
 
-        data = await from(data)
-            .pipe(skip((this._paging.page - 1) * (isNaN(this._paging.itemsPerPage) ? 0 : this._paging.itemsPerPage)),
-                  isNaN(this._paging.itemsPerPage) ? ((source: Observable<TData>) => source) : take(this._paging.itemsPerPage),
-                  toArray())
-            .toPromise();
+        data = await lastValueFrom(from(data)
+            .pipe(skip(((this.paging?.page ?? 1) - 1) * (isNaN(this.paging?.itemsPerPage ?? 0) ? 0 : (this.paging?.itemsPerPage ?? 0))),
+                  isNaN(this.paging?.itemsPerPage ?? 0) ? ((source: Observable<TData>) => source) : take(this.paging?.itemsPerPage ?? 0),
+                  toArray()));
 
-        this._state = (data && data.length) ? DataLoaderState.Loaded : DataLoaderState.NoData;
-        this.stateChange.emit();
+        this.ɵstate = (data && data.length) ? DataLoaderState.Loaded : DataLoaderState.NoData;
+        this.stateChangeSubject.next();
 
-        this._result =
+        this.ɵresult =
         {
             data: data,
-            totalCount: this._options.data.length
+            totalCount: this.ɵoptions.data.length
         };
 
-        this.resultChange.emit();
+        this.resultChangeSubject.next();
     }
 }