+++ src/plugins/dataLoader/sync/syncDataLoader.interface.ts
@@ -1,24 +1,24 @@
-import {DataLoaderOptions, DataLoader, DataResponse} from '../dataLoader.interface';
+import {DataLoader, DataLoaderOptions, DataResponse} from '../../../interfaces';
 
 /**
  * Sync data loader options
  */
-export interface SyncDataLoaderOptions<TData = any, TOrdering = any> extends DataLoaderOptions
+export interface SyncDataLoaderOptions<TData = unknown, TOrdering = unknown> extends DataLoaderOptions
 {
     /**
      * Static data which will be rendered by grid
      */
-    data?: TData[];
+    data: TData[];
 
     /**
      * Method used for ordering data
      */
-    orderData?: (data: TData[], ordering?: TOrdering) => TData[];
+    orderData: (data: TData[], ordering?: TOrdering) => TData[];
 }
 
 /**
  * Public API for 'SyncDataLoaderComponent'
  */
-export interface SyncDataLoader<TData = any> extends DataLoader<DataResponse<TData>>
+export interface SyncDataLoader<TData = unknown> extends DataLoader<DataResponse<TData>>
 {
 }